[WIP] Split layers in to own modules
