+++ qgreenland/tasks/main.py
@@ -10,11 +10,12 @@ from qgreenland.constants import (ASSETS_DIR,
                                   TMP_DIR,
                                   TaskType,
                                   ZIP_TRIGGERFILE)
-from qgreenland.tasks.layers import (ArcticDEM,
-                                     BedMachineDataset,
-                                     Coastlines,
-                                     GlacierTerminus)
-from qgreenland.util.misc import cleanup_intermediate_dirs
+from qgreenland.tasks.layers.arctic_dem import ArcticDEM
+from qgreenland.tasks.layers.bedmachine import BedMachineDataset
+from qgreenland.tasks.layers.coastlines import Coastlines
+from qgreenland.tasks.layers.glacier_terminus import GlacierTerminus
+from qgreenland.util.misc import (cleanup_intermediate_dirs,
+                                  get_layer_config)
 from qgreenland.util.qgis import make_qgis_project_file
 
 
@@ -41,6 +42,10 @@ class CreateProjectFile(luigi.Task):
         # iterates over the entire configuration file, so if you want to reduce
         # the stuff being run you have to edit many files, yuck. Even better,
         # make the config file drive the required layers here.
+        cfg = get_layer_config()
+        breakpoint()
+        print(cfg)
+
         yield ArcticDEM()
         yield BedMachineDataset('surface')
         yield BedMachineDataset('thickness')
